backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
python-multipart==0.0.9
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Shapely==2.0.4
geojson==3.1.0
python-dotenv==1.0.1
